2.7.6 INT-BITE

+++ ConsoleApp2/Program.cs
@@ -6,7 +6,7 @@ class MainClass
         Console.Write("Enter your name:");
         string name = Console.ReadLine();
         Console.Write("Enter your age:");
-        int age = int.Parse(Console.ReadLine());
+        byte age = (byte)int.Parse(Console.ReadLine());
         Console.WriteLine("Your name is {0} and age is {1}", name, age);
         Console.ReadLine();
     }
